Crate google_vault1

This documentation was generated from Vault crate version 1.0.9+20190614, where 20190614 is the exact revision of the vault:v1 schema built by the mako code generator v1.0.9.

Everything else about the Vault v1 API can be found at the official documentation site. The original source code is on github.

Features

Handle the following Resources with ease from the central hub ...

matters

add permissions, close, create, delete, exports create, exports delete, exports get, exports list, get, holds accounts create, holds accounts delete, holds accounts list, holds add held accounts, holds create, holds delete, holds get, holds list, holds remove held accounts, holds update, list, remove permissions, reopen, saved queries create, saved queries delete, saved queries get, saved queries list, undelete and update

Not what you are looking for ? Find all other Google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

creates Method Builders which in turn allow access to individual Call Builders

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

All structures are marked with applicable traits to further categorize them and ease browsing.

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.matters().holds_delete(...).doit()
let r = hub.matters().holds_list(...).doit()
let r = hub.matters().saved_queries_delete(...).doit()
let r = hub.matters().exports_create(...).doit()
let r = hub.matters().update(...).doit()
let r = hub.matters().holds_get(...).doit()
let r = hub.matters().holds_accounts_list(...).doit()
let r = hub.matters().exports_list(...).doit()
let r = hub.matters().exports_get(...).doit()
let r = hub.matters().holds_add_held_accounts(...).doit()
let r = hub.matters().undelete(...).doit()
let r = hub.matters().remove_permissions(...).doit()
let r = hub.matters().saved_queries_list(...).doit()
let r = hub.matters().add_permissions(...).doit()
let r = hub.matters().holds_update(...).doit()
let r = hub.matters().close(...).doit()
let r = hub.matters().get(...).doit()
let r = hub.matters().holds_create(...).doit()
let r = hub.matters().holds_remove_held_accounts(...).doit()
let r = hub.matters().create(...).doit()
let r = hub.matters().list(...).doit()
let r = hub.matters().reopen(...).doit()
let r = hub.matters().holds_accounts_delete(...).doit()
let r = hub.matters().exports_delete(...).doit()
let r = hub.matters().holds_accounts_create(...).doit()
let r = hub.matters().saved_queries_create(...).doit()
let r = hub.matters().delete(...).doit()
let r = hub.matters().saved_queries_get(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
google-vault1 = "*"
# This project intentionally uses an old version of Hyper. See
# https://github.com/Byron/google-apis-rs/issues/173 for more
# information.
hyper = "^0.10"
hyper-rustls = "^0.6"
serde = "^1.0"
serde_json = "^1.0"
yup-oauth2 = "^1.0"

A complete example

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_vault1 as vault1;
use vault1::{Result, Error};
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use vault1::Vault;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Vault::new(hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())), auth);
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.matters().holds_list("matterId")
             .view("dolores")
             .page_token("kasd")
             .page_size(-22)
             .doit();
 
match result {
    Err(e) => match e {
        // The Error enum provides details about what exactly happened.
        // You can also just use its `Debug`, `Display` or `Error` traits
         Error::HttpError(_)
        |Error::MissingAPIKey
        |Error::MissingToken(_)
        |Error::Cancelled
        |Error::UploadSizeLimitExceeded(_, _)
        |Error::Failure(_)
        |Error::BadRequest(_)
        |Error::FieldClash(_)
        |Error::JsonDecodeError(_, _) => println!("{}", e),
    },
    Ok(res) => println!("Success: {:?}", res),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

Uploads and Downloads

If a method supports downloads, the response body, which is part of the Result, should be read by you to obtain the media. If such a method also supports a Response Result, it will return that by default. You can see it as meta-data for the actual media. To trigger a media download, you will have to set up the builder by making this call: .param("alt", "media").

Methods supporting uploads can do so using up to 2 different protocols: simple and resumable. The distinctiveness of each is represented by customized doit(...) methods, which are then named upload(...) and upload_resumable(...) respectively.

Customization and Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

Optional Parts in Server-Requests

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifiable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are moved

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Structs

AccountInfo
Accounts to search

AddHeldAccountResult
A status detailing the status of each account creation, and the HeldAccount, if successful.

AddHeldAccountsRequest
Add a list of accounts to a hold.

AddHeldAccountsResponse
Response for batch create held accounts.

AddMatterPermissionsRequest
Add an account with the permission specified. The role cannot be owner. If an account already has a role in the matter, it will be overwritten.

Chunk

CloseMatterRequest
Close a matter by ID.

CloseMatterResponse
Response to a CloseMatterRequest.

CloudStorageFile
An export file on cloud storage

CloudStorageSink
Export sink for cloud storage files.

ContentRange
Implements the Content-Range header, for serialization only

CorpusQuery
Corpus specific queries.

DefaultDelegate
A delegate with a conservative default implementation, which is used if no other delegate is set.

DriveExportOptions
The options for Drive export.

DriveOptions
Drive search advanced options

DummyNetworkStream

Empty
A generic empty message that you can re-use to avoid defining duplicated empty messages in your APIs. A typical example is to use it as the request or the response type of an API method. For instance:

ErrorResponse
A utility to represent detailed errors we might see in case there are BadRequests. The latter happen if the sent parameters or request structures are unsound

Export
An export

ExportOptions
Export advanced options

ExportStats
Stats of an export.

GroupsExportOptions
The options for groups export.

HangoutsChatExportOptions
The options for hangouts chat export.

HangoutsChatInfo
Accounts to search

HangoutsChatOptions
Hangouts chat search advanced options

HeldAccount
An account being held in a particular hold. This structure is immutable. This can be either a single user or a google group, depending on the corpus.

HeldDriveQuery
Query options for Drive holds.

HeldGroupsQuery
Query options for group holds.

HeldHangoutsChatQuery
Query options for hangouts chat holds.

HeldMailQuery
Query options for mail holds.

HeldOrgUnit
A organizational unit being held in a particular hold. This structure is immutable.

Hold
Represents a hold within Vault. A hold restricts purging of artifacts based on the combination of the query and accounts restrictions. A hold can be configured to either apply to an explicitly configured set of accounts, or can be applied to all members of an organizational unit.

JsonServerError
A utility type which can decode a server response that indicates error

ListExportsResponse
The holds for a matter.

ListHeldAccountsResponse
Returns a list of held accounts for a hold.

ListHoldsResponse
The holds for a matter.

ListMattersResponse
Provides the list of matters.

ListSavedQueriesResponse
Definition of the response for method ListSaveQuery.

MailExportOptions
The options for mail export.

MailOptions
Mail search advanced options

Matter
Represents a matter.

MatterAddPermissionCall
Adds an account as a matter collaborator.

MatterCloseCall
Closes the specified matter. Returns matter with updated state.

MatterCreateCall
Creates a new matter with the given name and description. The initial state is open, and the owner is the method caller. Returns the created matter with default view.

MatterDeleteCall
Deletes the specified matter. Returns matter with updated state.

MatterExportCreateCall
Creates an Export.

MatterExportDeleteCall
Deletes an Export.

MatterExportGetCall
Gets an Export.

MatterExportListCall
Lists Exports.

MatterGetCall
Gets the specified matter.

MatterHoldAccountCreateCall
Adds a HeldAccount to a hold. Accounts can only be added to a hold that has no held_org_unit set. Attempting to add an account to an OU-based hold will result in an error.

MatterHoldAccountDeleteCall
Removes a HeldAccount from a hold. If this request leaves the hold with no held accounts, the hold will not apply to any accounts.

MatterHoldAccountListCall
Lists HeldAccounts for a hold. This will only list individually specified held accounts. If the hold is on an OU, then use Admin SDK to enumerate its members.

MatterHoldAddHeldAccountCall
Adds HeldAccounts to a hold. Returns a list of accounts that have been successfully added. Accounts can only be added to an existing account-based hold.

MatterHoldCreateCall
Creates a hold in the given matter.

MatterHoldDeleteCall
Removes a hold by ID. This will release any HeldAccounts on this Hold.

MatterHoldGetCall
Gets a hold by ID.

MatterHoldListCall
Lists holds within a matter. An empty page token in ListHoldsResponse denotes no more holds to list.

MatterHoldRemoveHeldAccountCall
Removes HeldAccounts from a hold. Returns a list of statuses in the same order as the request. If this request leaves the hold with no held accounts, the hold will not apply to any accounts.

MatterHoldUpdateCall
Updates the OU and/or query parameters of a hold. You cannot add accounts to a hold that covers an OU, nor can you add OUs to a hold that covers individual accounts. Accounts listed in the hold will be ignored.

MatterListCall
Lists matters the user has access to.

MatterMethods
A builder providing access to all methods supported on matter resources. It is not used directly, but through the Vault hub.

MatterPermission
Currently each matter only has one owner, and all others are collaborators. When an account is purged, its corresponding MatterPermission resources cease to exist.

MatterRemovePermissionCall
Removes an account as a matter collaborator.

MatterReopenCall
Reopens the specified matter. Returns matter with updated state.

MatterSavedQueryCreateCall
Creates a saved query.

MatterSavedQueryDeleteCall
Deletes a saved query by Id.

MatterSavedQueryGetCall
Retrieves a saved query by Id.

MatterSavedQueryListCall
Lists saved queries within a matter. An empty page token in ListSavedQueriesResponse denotes no more saved queries to list.

MatterUndeleteCall
Undeletes the specified matter. Returns matter with updated state.

MatterUpdateCall
Updates the specified matter. This updates only the name and description of the matter, identified by matter id. Changes to any other fields are ignored. Returns the default view of the matter.

MethodInfo
Contains information about an API request.

MultiPartReader
Provides a Read interface that converts multiple parts into the protocol identified by RFC2387. Note: This implementation is just as rich as it needs to be to perform uploads to google APIs, and might not be a fully-featured implementation.

OrgUnitInfo
Org Unit to search

Query
A query definition relevant for search & export.

RangeResponseHeader

RemoveHeldAccountsRequest
Remove a list of accounts from a hold.

RemoveHeldAccountsResponse
Response for batch delete held accounts.

RemoveMatterPermissionsRequest
Remove an account as a matter collaborator.

ReopenMatterRequest
Reopen a matter by ID.

ReopenMatterResponse
Response to a ReopenMatterRequest.

ResumableUploadHelper
A utility type to perform a resumable upload from start to end.

SavedQuery
Definition of the saved query.

ServerError

ServerMessage

SharedDriveInfo
Shared drives to search

Status
The Status type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by gRPC. Each Status message contains three pieces of data: error code, error message, and error details.

TeamDriveInfo
Team Drives to search

UndeleteMatterRequest
Undelete a matter by ID.

UserInfo
User's information.

Vault
Central instance to access all Vault related resource activities

XUploadContentType
The X-Upload-Content-Type header.

Enums

Error

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.

Traits

CallBuilder
Identifies types which represent builders for a particular resource method

Delegate
A trait specifying functionality to help controlling any request performed by the API. The trait has a conservative default implementation.

Hub
Identifies the Hub. There is only one per library, this trait is supposed to make intended use more explicit. The hub allows to access all resource methods more easily.

MethodsBuilder
Identifies types for building methods of a particular resource type

NestedType
Identifies types which are only used by other types internally. They have no special meaning, this trait just marks them for completeness.

Part
Identifies types which are only used as part of other types, which usually are carrying the Resource trait.

ReadSeek
A utility to specify reader types which provide seeking capabilities too

RequestValue
Identifies types which are used in API requests.

Resource
Identifies types which can be inserted and deleted. Types with this trait are most commonly used by clients of this API.

ResponseResult
Identifies types which are used in API responses.

ToParts
A trait for all types that can convert themselves into a parts string

UnusedType
Identifies types which are not actually used by the API This might be a bug within the google API schema.

Functions

remove_json_null_values

Type Definitions

Result
A universal result type used as return for all calls.