Struct google_vault1::MatterMethods

pub struct MatterMethods<'a, C, A> where
    C: 'a,
    A: 'a,  { /* fields omitted */ }

A builder providing access to all methods supported on matter resources. It is not used directly, but through the Vault hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_vault1 as vault1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use vault1::Vault;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Vault::new(hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `add_permissions(...)`, `close(...)`, `create(...)`, `delete(...)`, `exports_create(...)`, `exports_delete(...)`, `exports_get(...)`, `exports_list(...)`, `get(...)`, `holds_accounts_create(...)`, `holds_accounts_delete(...)`, `holds_accounts_list(...)`, `holds_add_held_accounts(...)`, `holds_create(...)`, `holds_delete(...)`, `holds_get(...)`, `holds_list(...)`, `holds_remove_held_accounts(...)`, `holds_update(...)`, `list(...)`, `remove_permissions(...)`, `reopen(...)`, `saved_queries_create(...)`, `saved_queries_delete(...)`, `saved_queries_get(...)`, `saved_queries_list(...)`, `undelete(...)` and `update(...)`
// to build up your call.
let rb = hub.matters();

Methods

impl<'a, C, A> MatterMethods<'a, C, A>

pub fn holds_delete(
&self,
matter_id: &str,
hold_id: &str
) -> MatterHoldDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Removes a hold by ID. This will release any HeldAccounts on this Hold.

Arguments

matterId - The matter ID.

holdId - The hold ID.

pub fn holds_list(&self, matter_id: &str) -> MatterHoldListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists holds within a matter. An empty page token in ListHoldsResponse denotes no more holds to list.

Arguments

matterId - The matter ID.

pub fn saved_queries_delete(
&self,
matter_id: &str,
saved_query_id: &str
) -> MatterSavedQueryDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes a saved query by Id.

Arguments

matterId - The matter id of the parent matter for which the saved query is to be deleted.

savedQueryId - Id of the saved query to be deleted.

pub fn exports_create(
&self,
request: Export,
matter_id: &str
) -> MatterExportCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates an Export.

Arguments

request - No description provided.

matterId - The matter ID.

pub fn update(
&self,
request: Matter,
matter_id: &str
) -> MatterUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Updates the specified matter. This updates only the name and description of the matter, identified by matter id. Changes to any other fields are ignored. Returns the default view of the matter.

Arguments

request - No description provided.

matterId - The matter ID.

pub fn holds_get(
&self,
matter_id: &str,
hold_id: &str
) -> MatterHoldGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets a hold by ID.

Arguments

matterId - The matter ID.

holdId - The hold ID.

pub fn holds_accounts_list(
&self,
matter_id: &str,
hold_id: &str
) -> MatterHoldAccountListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists HeldAccounts for a hold. This will only list individually specified held accounts. If the hold is on an OU, then use Admin SDK to enumerate its members.

Arguments

matterId - The matter ID.

holdId - The hold ID.

pub fn exports_list(&self, matter_id: &str) -> MatterExportListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists Exports.

Arguments

matterId - The matter ID.

pub fn exports_get(
&self,
matter_id: &str,
export_id: &str
) -> MatterExportGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets an Export.

Arguments

matterId - The matter ID.

exportId - The export ID.

pub fn holds_add_held_accounts(
&self,
request: AddHeldAccountsRequest,
matter_id: &str,
hold_id: &str
) -> MatterHoldAddHeldAccountCall<'a, C, A>

Create a builder to help you perform the following task:

Adds HeldAccounts to a hold. Returns a list of accounts that have been successfully added. Accounts can only be added to an existing account-based hold.

Arguments

request - No description provided.

matterId - The matter ID.

holdId - The hold ID.

pub fn undelete(
&self,
request: UndeleteMatterRequest,
matter_id: &str
) -> MatterUndeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Undeletes the specified matter. Returns matter with updated state.

Arguments

request - No description provided.

matterId - The matter ID.

pub fn remove_permissions(
&self,
request: RemoveMatterPermissionsRequest,
matter_id: &str
) -> MatterRemovePermissionCall<'a, C, A>

Create a builder to help you perform the following task:

Removes an account as a matter collaborator.

Arguments

request - No description provided.

matterId - The matter ID.

pub fn saved_queries_list(
&self,
matter_id: &str
) -> MatterSavedQueryListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists saved queries within a matter. An empty page token in ListSavedQueriesResponse denotes no more saved queries to list.

Arguments

matterId - The matter id of the parent matter for which the saved queries are to be retrieved.

pub fn add_permissions(
&self,
request: AddMatterPermissionsRequest,
matter_id: &str
) -> MatterAddPermissionCall<'a, C, A>

Create a builder to help you perform the following task:

Adds an account as a matter collaborator.

Arguments

request - No description provided.

matterId - The matter ID.

pub fn holds_update(
&self,
request: Hold,
matter_id: &str,
hold_id: &str
) -> MatterHoldUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Updates the OU and/or query parameters of a hold. You cannot add accounts to a hold that covers an OU, nor can you add OUs to a hold that covers individual accounts. Accounts listed in the hold will be ignored.

Arguments

request - No description provided.

matterId - The matter ID.

holdId - The ID of the hold.

pub fn close(
&self,
request: CloseMatterRequest,
matter_id: &str
) -> MatterCloseCall<'a, C, A>

Create a builder to help you perform the following task:

Closes the specified matter. Returns matter with updated state.

Arguments

request - No description provided.

matterId - The matter ID.

pub fn get(&self, matter_id: &str) -> MatterGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets the specified matter.

Arguments

matterId - The matter ID.

pub fn holds_create(
&self,
request: Hold,
matter_id: &str
) -> MatterHoldCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a hold in the given matter.

Arguments

request - No description provided.

matterId - The matter ID.

pub fn holds_remove_held_accounts(
&self,
request: RemoveHeldAccountsRequest,
matter_id: &str,
hold_id: &str
) -> MatterHoldRemoveHeldAccountCall<'a, C, A>

Create a builder to help you perform the following task:

Removes HeldAccounts from a hold. Returns a list of statuses in the same order as the request. If this request leaves the hold with no held accounts, the hold will not apply to any accounts.

Arguments

request - No description provided.

matterId - The matter ID.

holdId - The hold ID.

pub fn create(&self, request: Matter) -> MatterCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a new matter with the given name and description. The initial state is open, and the owner is the method caller. Returns the created matter with default view.

Arguments

request - No description provided.

pub fn list(&self) -> MatterListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists matters the user has access to.

pub fn reopen(
&self,
request: ReopenMatterRequest,
matter_id: &str
) -> MatterReopenCall<'a, C, A>

Create a builder to help you perform the following task:

Reopens the specified matter. Returns matter with updated state.

Arguments

request - No description provided.

matterId - The matter ID.

pub fn holds_accounts_delete(
&self,
matter_id: &str,
hold_id: &str,
account_id: &str
) -> MatterHoldAccountDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Removes a HeldAccount from a hold. If this request leaves the hold with no held accounts, the hold will not apply to any accounts.

Arguments

matterId - The matter ID.

holdId - The hold ID.

accountId - The ID of the account to remove from the hold.

pub fn exports_delete(
&self,
matter_id: &str,
export_id: &str
) -> MatterExportDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes an Export.

Arguments

matterId - The matter ID.

exportId - The export ID.

pub fn holds_accounts_create(
&self,
request: HeldAccount,
matter_id: &str,
hold_id: &str
) -> MatterHoldAccountCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Adds a HeldAccount to a hold. Accounts can only be added to a hold that has no held_org_unit set. Attempting to add an account to an OU-based hold will result in an error.

Arguments

request - No description provided.

matterId - The matter ID.

holdId - The hold ID.

pub fn saved_queries_create(
&self,
request: SavedQuery,
matter_id: &str
) -> MatterSavedQueryCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a saved query.

Arguments

request - No description provided.

matterId - The matter id of the parent matter for which the saved query is to be created.

pub fn delete(&self, matter_id: &str) -> MatterDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes the specified matter. Returns matter with updated state.

Arguments

matterId - The matter ID

pub fn saved_queries_get(
&self,
matter_id: &str,
saved_query_id: &str
) -> MatterSavedQueryGetCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves a saved query by Id.

Arguments

matterId - The matter id of the parent matter for which the saved query is to be retrieved.

savedQueryId - Id of the saved query to be retrieved.